refactor(lotto): LottoNumberGenerator 가 하던 책임을 LottoNumber 으로 위임
