+++ src/main/java/lotto/domain/LottoNumber.java
@@ -1,10 +1,22 @@
 package lotto.domain;
 
+import java.util.ArrayList;
+import java.util.Collections;
+import java.util.List;
 import java.util.Objects;
+import java.util.stream.Collectors;
+import java.util.stream.IntStream;
 
 public class LottoNumber implements Comparable<LottoNumber> {
-    public static final int MIN_NUMBER = 1;
-    public static final int MAX_NUMBER = 45;
+    private static final int MIN_NUMBER = 1;
+    private static final int MAX_NUMBER = 45;
+    private static final List<LottoNumber> ALL_LOTTO_NUMBERS;
+
+    static {
+        ALL_LOTTO_NUMBERS = IntStream.rangeClosed(LottoNumber.MIN_NUMBER, LottoNumber.MAX_NUMBER)
+            .mapToObj(LottoNumber::new)
+            .collect(Collectors.toUnmodifiableList());
+    }
 
     private final int number;
 
@@ -14,6 +26,26 @@ public class LottoNumber implements Comparable<LottoNumber> {
         }
 
         this.number = number;
+    }
+
+    public static List<LottoNumber> getNewLottoNumbers(final int generateCount) {
+        if (ALL_LOTTO_NUMBERS.size() < generateCount) {
+            throw new IllegalArgumentException("전체 로또 개수를 초과하여 생성 할 수 없습니다.");
+        }
+
+        final List<LottoNumber> pickedNumbers = pickLottoNumbers(shuffle(), generateCount);
+        Collections.sort(pickedNumbers);
+        return Collections.unmodifiableList(pickedNumbers);
+    }
+
+    private static List<LottoNumber> pickLottoNumbers(final List<LottoNumber> shuffledLottoNumbers, final int lottoPickCount) {
+        return shuffledLottoNumbers.subList(0, lottoPickCount);
+    }
+
+    private static List<LottoNumber> shuffle() {
+        final List<LottoNumber> shuffledLottoNumbers = new ArrayList<>(ALL_LOTTO_NUMBERS);
+        Collections.shuffle(shuffledLottoNumbers);
+        return shuffledLottoNumbers;
     }
 
     @Override
