+++ src/test/java/lotto/domain/LottoRankGroupTest.java
@@ -1,9 +1,9 @@
 package lotto.domain;
 
-import lotto.util.LottoNumberGenerator;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
+import static lotto.domain.Lotto.LOTTO_NUMBER_COUNT;
 import static org.assertj.core.api.Assertions.assertThat;
 
 class LottoRankGroupTest {
@@ -12,7 +12,7 @@ class LottoRankGroupTest {
     void shouldAddLottoBundle() {
         final LottoRankGroup group = new LottoRankGroup();
 
-        group.addLotto(LottoRank.FIRST, new Lotto(LottoNumberGenerator.shuffleAndPickLottoNumbers()));
+        group.addLotto(LottoRank.FIRST, new Lotto(LottoNumber.getNewLottoNumbers(LOTTO_NUMBER_COUNT)));
 
         assertThat(group.getLottoBundleByRank(LottoRank.FIRST)).hasSize(1);
     }
